pyproject.toml
[tool.poetry]
name = "discord-youtube-notifier"
version = "0.1.0"
description = ""
authors = ["Suzuka Asagiri <admin@suzutan.jp>"]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.1.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
